[refactor](file-system)(step-2) remove env, file_utils and filesystem_utils (#18009)
Follow #17586.
This PR mainly changes:

Remove env/
Remove FileUtils/FilesystemUtils
Some methods are moved to LocalFileSystem
Remove olap/file_cache
Add s3 client cache for s3 file system
In my test, the time of open s3 file can be reduced significantly
Fix cold/hot separation bug for s3 fs.
This is the last PR of #17764.
After this, all IO operation should be in io/fs.

Except for tests in #17586, I also tested some case related to fs io:

clone
concurrency query on local/s3/hdfs
load error log create and clean
disk metrics

+++ be/src/io/file_factory.h
@@ -40,7 +40,7 @@ struct FileSystemProperties {
 struct FileDescription {
     std::string path;
     int64_t start_offset;
-    size_t file_size;
+    int64_t file_size;
 };
 
 class FileFactory {
